Add files via upload
Develop a hospital software with 2 services. These services are Service_1 and Service_2 respectively.
Service_1: There should be an unlimited number of patient admissions. Depending on the urgency, the 
patient who comes later should be treated earlier. Patients should be able to be selected or removed 
from the patient list by name or order
Service_2: There should be an unlimited number of patient admissions. In this service, the first patient 
will be treated first. It should work with the First in First Out (FIFO) principle.